pyproject.toml
[project]
name = "database-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphviz>=0.21",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.0.10",
    "matplotlib>=3.10.8",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
]
